fix: some fixes to json tree and flow
